Remove redundant specifyLogFile Parameter and Unity process monitor
The log file name can be used to decide if one is created. If the name is empty, no log file will be created. If a name is specified, by default it is, a log file is created.
Removed the Unity Process monitor since this approach has too many flaws. It's time to find a way to do this using the command alone.