Correcting UnitySetupTask postjobexecution repetition (#292)
* Move input variables to own script on UnitySetup

When using the UnitySetup task, I noticed that the post-job always tries to reinstall unity again, even though it should just check for the license deactivation, as the unity-setup.post script instructs.
The cause seems to be that the unity-editor-deactivation script imports variables from the main unity-setup script, which in turn runs that whole script again. While this did not impede any functions, it was an additional 2 minutes of time the agent spent checking the Unity installation, instead of just logging out.
Moving all variables to a seperate file with no function call fixed the issue for me.

* Implementing installEditor bool

While checking for the post-job repetition issue, I noticed that the installEditor bool is not actually used anywhere, and added it to the install scripts.

* Moved deactivateSeatOnComplete bool

Sometimes a pipeline would throw an error during the post-job due to "getProjectUnityVersion()" failing for example, even though deactivationOnComplete was turned off anyway and the project version was not needed. Checking the bool right at the beginning of the script avoids unneccessary work.

* Fixed typo

:)